+++ src/main/java/com/codenear/butterfly/kakaoPay/application/SinglePaymentService.java
@@ -9,11 +9,11 @@ import com.codenear.butterfly.kakaoPay.domain.OrderDetails;
 import com.codenear.butterfly.kakaoPay.domain.SinglePayment;
 import com.codenear.butterfly.kakaoPay.domain.dto.OrderType;
 import com.codenear.butterfly.kakaoPay.domain.dto.PaymentStatus;
-import com.codenear.butterfly.kakaoPay.domain.dto.kakao.ApproveFreePaymentHandler;
-import com.codenear.butterfly.kakaoPay.domain.dto.kakao.ApproveHandler;
-import com.codenear.butterfly.kakaoPay.domain.dto.kakao.ApprovePaymentHandler;
 import com.codenear.butterfly.kakaoPay.domain.dto.kakao.ApproveResponseDTO;
 import com.codenear.butterfly.kakaoPay.domain.dto.kakao.ReadyResponseDTO;
+import com.codenear.butterfly.kakaoPay.domain.dto.kakao.handler.ApproveFreePaymentHandler;
+import com.codenear.butterfly.kakaoPay.domain.dto.kakao.handler.ApproveHandler;
+import com.codenear.butterfly.kakaoPay.domain.dto.kakao.handler.ApprovePaymentHandler;
 import com.codenear.butterfly.kakaoPay.domain.dto.order.OrderDTO;
 import com.codenear.butterfly.kakaoPay.domain.dto.rabbitmq.InventoryDecreaseMessageDTO;
 import com.codenear.butterfly.kakaoPay.domain.dto.request.BasePaymentRequestDTO;
